rename directory to expected python module heiarchy
