+++ .github/workflows/python_tests.yml
@@ -12,29 +12,28 @@ jobs:
       matrix:
         python-version: [3.11]
     steps:
-    - uses: actions/checkout@v4
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v5
-      with:
-        python-version: ${{ matrix.python-version }}
-        cache: 'pip'
-    - name: Install dependencies
-      run: |
-        sudo apt update
-        sudo apt install -y libasound2-dev qt6-base-dev qt6-tools-dev-tools
-        pip install --upgrade pip
-        pip install --upgrade setuptools
-        pip install flake8 pytest
-        pip install --ignore-installed -r requirements.txt
-        bash setup-environment.ps1
-    - name: Lint with flake8
-      run: |
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 src/ --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings.
-        flake8 src/ --count --exit-zero --statistics
-    - name: Test with pytest
-      run: pytest -v
-    - name: Typecheck with mypy
-      run: mypy src/main/python
-      
+      - uses: actions/checkout@v4
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+          cache: "pip"
+      - name: Install dependencies
+        run: |
+          sudo apt update
+          sudo apt install -y libasound2-dev qt6-base-dev qt6-tools-dev-tools
+          pip install --upgrade pip
+          pip install --upgrade setuptools
+          pip install flake8 pytest
+          pip install --ignore-installed -r requirements.txt
+          bash setup-environment.ps1
+      - name: Lint with flake8
+        run: |
+          # stop the build if there are Python syntax errors or undefined names
+          flake8 src/ --count --select=E9,F63,F7,F82 --show-source --statistics
+          # exit-zero treats all errors as warnings.
+          flake8 src/ --count --exit-zero --statistics
+      - name: Test with pytest
+        run: pytest -v
+      - name: Typecheck with mypy
+        run: mypy src/main/main
