+++ Doxyfile
@@ -51,7 +51,7 @@ PROJECT_BRIEF          = AYAB desktop software
 # pixels and the maximum width should not exceed 200 pixels. Doxygen will copy
 # the logo to the output directory.
 
-PROJECT_LOGO           = src/main/python/ayab/ayab_logo.jpg
+PROJECT_LOGO           = src/main/main/ayab/ayab_logo.jpg
 
 # The OUTPUT_DIRECTORY tag is used to specify the (relative or absolute) path
 # into which the generated documentation will be written. If a relative path is
