+++ src/main/main/ayab/ayab.py
@@ -47,7 +47,7 @@ from .engine.engine_fsm import Operation
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from main import AppContext
+    from ..main import AppContext
 
 
 class GuiMain(QMainWindow):
