+++ src/main/main/ayab/ayab_logo_rc.qrc
