+++ src/main/main/ayab/engine/garamond-lowercase-e.png
