+++ src/main/main/ayab/engine/lowercase_e_rc.qrc
