+++ src/main/main/ayab/engine/lowercase_e_reversed_rc.qrc
