+++ src/main/main/ayab/engine/pattern.py
