+++ src/main/main/ayab/tests/README.md
@@ -1,6 +1,6 @@
 # To run tests
 
-From `src/main/python`
+From `src/main/main`
 
 Run:
